Established constraints on Attribute values
Constraint_Type<T> is a template that constraints with types specific
interfaces will be implemented. So far ElementID is implemented and does
what we want - the editor can only set the Attribute value to that of an
Element of the types specified in the definition. Definitions are going
to need cleaning up, constructing them is a bit messy right now.