Unfinished stuff
This branch is probably a dead end now. The idea was to split data into
the smallest chunks that will be relevant to Qt displaying it, and it
could probably work, it just got incredibly messy to make the data
structure fit how the Qt model system would display it.