More template madness
So the whole 'forward everything to the manager' ground to a halt and I
ended up doing away with it and organising the handles how I was going
to before. Going to need to figure out where to do the part that will be
responsible for keeping Element names unique.