Project signals
Now have some signals for File changes and rudimentary File saving and
loading. Having things in the outliner move after renaming them is a
pain in the ass and I'm not sure what that does that causes it to break.
Might ditch that and figure out drag and drop moving?