Seemingly functional outliner and editors
The outliner seems to be doing most of what I want! The selection even
moves around properly. There is also the ability to open/close editors
and the data all stays in sync. It's all a bit of a mess though, things
need to be sectioned off - like the outliner and tabwidget updates.