Data sectioning
Starting to split the data up so that coding repeats can be reduced when
we get to multiple component types with shared implementation details.
Right now in the midst of figuring out what signals go into and come out
of the Project. Might need to implement undoing to figure this out since
that will a lot of signalling as it is used.