Rearranging and undo stuff
To make the signals in and out a bit easier to understand some things
have been moved around and there are template instantiations. There's
also an undo system, and it works, but it is currently out of sync.
There's an order of update issue with it. Going to need that data class
implementation to stay sane when we get to defining lots of things to
edit...