Refactoring
Lots of stuff has been moved around. project::object is now responsible
for the entire command system, meaning that it's hidden from view. There
is a standardised way of accessing the unambiguous project::object
interface. Next step is to make another type like file and see if this
system is actually expandable.