Widget access reorganisation test
Moved some stuff around in order to test a different way of getting to
widget traits since things can have the same value type but different
widget behaviour. A bit messy but it works. Going to need to figure out
all individual behaviour for a given member. and how to declare that.
This is all coing down to how to split off behaviour for different
things in the templates.