Cleanup and editing changes
So it turns out you can leave a QLineEdit that has a QValidator while
the text is not valid, and there's no signal for when that happens. I
built a way of adding one but it has it's own complications. With
tweaking it could be made less of an arse.