Yet more cleanup and testing
Project data is getting there. Having decided to change the object name
setup so that they're unique across all managed types there are some
more bits that need to be built.  Also we ahve this annoying thing going
on where it might be simpler to just build the project data imp class
using the project interface class, but doing so makes testing harder to
do, and that is proving very useful.