Project implementation fiddling
Split out the signalbox logic into it's own object so that seperate
parts are seperate. Some experiments into changing the inheritance
chains so that virtual inheritance can be eliminated since it looks
pretty bloated. It can be removed, it's just going to be messy in source
code rather than the compiled code...