Signalbox reworking
I realised that the part that manages data was actually not all that
well defined and would get a bit weird with undoing, so I went back and
looked over what the data setup should actually be and determined the
required signals for it and rebuilt the signalbox system. Now I need to
decide how I'm going to build the part the manages the raw data. Since
it's going to be used by another level to add the signalling and history
tracking the functions are going to need to be setup to give information
about operations that can fail.