Entity_Manager testing
Setting up testing stuff. Things seem to be working so far. Might want
to pin down the name fixing logic - it works in that it makes the names
unique but there are some cases where it behaves a little wierd. Could
probably rebuild the GUI with what there is so far.