Establishing member data types
Started work on the entity member data system and made strings in the
system use a library specific implementation so that they can be tweaked
later and use copy on write. This is mostly because it seems like
there's a lot of string copying without it, but I don't know if that's
really worth bothering with. It's done now. Also established the variant
and filepath types, which could both do with going pimpl because the
things they need aren't available in the standard library yet.